Code/TestApp/requirements.txt
PyQt6
pandas
openpyxl
python-docx
pymupdf